requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.7
distro==1.9.0
Flask==3.1.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.9.0
Mako==1.3.8
MarkupSafe==3.0.2
numpy==2.2.3
openai==1.66.3
opencv-python==4.11.0.86
packaging==24.2
pdf2image==1.17.0
pillow==11.1.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pytesseract==0.3.13
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
